Converted Search Bar and Filter to components
